requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
PyJWT
pytest
httpx
locust